+++ src/test/java/io/shiftleft/fuzzyc2cpg/antlrparsers/functionparser/AssignmentTests.java
@@ -3,7 +3,7 @@ package io.shiftleft.fuzzyc2cpg.antlrparsers.functionparser;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import io.shiftleft.fuzzyc2cpg.parser.AntlrParserDriver;
+import io.shiftleft.fuzzyc2cpg.passes.astcreation.AntlrParserDriver;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.junit.Test;
