+++ src/test/java/io/shiftleft/fuzzyc2cpg/antlrparsers/functionparser/FunctionParserTestBase.java
@@ -1,6 +1,7 @@
 package io.shiftleft.fuzzyc2cpg.antlrparsers.functionparser;
 
 
+import io.shiftleft.fuzzyc2cpg.passes.astcreation.AntlrParserDriver;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
@@ -8,8 +9,6 @@ import org.antlr.v4.runtime.Token;
 
 import io.shiftleft.fuzzyc2cpg.FunctionLexer;
 import io.shiftleft.fuzzyc2cpg.FunctionParser;
-import io.shiftleft.fuzzyc2cpg.parser.AntlrParserDriver;
-import io.shiftleft.fuzzyc2cpg.parser.TokenSubStream;
 import io.shiftleft.fuzzyc2cpg.parser.functions.AntlrCFunctionParserDriver;
 
 public class FunctionParserTestBase {
